src/accounts/serializers.py
from rest_framework import serializers

from src.accounts.models import Account


class AccountSerializer(serializers.ModelSerializer):

    class Meta:
        model = Account
        fields = (
            'id',
            'username',
            'first_name',
            'last_name',
            'profile_picture',
        )
        read_only_fields = ('username', )


class CreateAccountSerializer(serializers.ModelSerializer):
    tokens = serializers.SerializerMethodField()

    def get_tokens(self, user):
        return user.get_tokens()

    def create(self, validated_data):
        # call create_user on user object. Without this
        # the password will be stored in plain text.
        user = Account.objects.create_user(**validated_data)
        return user

    class Meta:
        model = Account
        fields = (
            'id',
            'username',
            'password',
            'first_name',
            'last_name',
            'email',
            'tokens',
        )
        read_only_fields = ('tokens', )
        extra_kwargs = {'password': {'write_only': True}}
